requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
networkx==3.2.1
python-multipart==0.0.6
tqdm==4.66.1
python-dotenv==1.0.0

